+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleLanguage.java
@@ -123,6 +123,25 @@ public abstract class TruffleLanguage<C> {
          * @return array of MIME types assigned to your language files
          */
         String[] mimeType();
+
+        /**
+         * Indicates whether the language implements an interactive response to evaluation of
+         * {@link Source#isInteractive() interactive sources}. When
+         * {@link com.oracle.truffle.api.vm.PolyglotEngine#eval} is called over an
+         * {@link Source#isInteractive() interactive source}, interactive languages print the result
+         * to {@link com.oracle.truffle.api.vm.PolyglotEngine.Builder#setOut(OutputStream) standard
+         * output} or {@link com.oracle.truffle.api.vm.PolyglotEngine.Builder#setErr(OutputStream)
+         * error output} and can use
+         * {@link com.oracle.truffle.api.vm.PolyglotEngine.Builder#setIn(InputStream) standard
+         * input}. Non-interactive languages are not expected to use the polyglot engine streams in
+         * an interactive way and whoever supplies an interactive source is expected to implement an
+         * appropriate response.
+         *
+         * @return <code>true</code> if the language implements an interactive response to
+         *         evaluation of interactive sources.
+         * @since 0.22
+         */
+        boolean interactive() default false;
     }
 
     /**